Minimize API calls for the cost.
